backend/requirements.txt
# Core Framework
Flask==3.0.0
python-dotenv==1.0.0
Flask-CORS==4.0.0

# AI & LLM
google-generativeai==0.3.2
crewai==0.28.8
langchain>=0.1.10,<0.2.0
langchain-google-genai>=0.0.6

# Database & ORM
SQLAlchemy>=2.0.27,<3.0.0
Flask-SQLAlchemy==3.1.1
alembic==1.13.1
psycopg2-binary==2.9.9

# WebSockets & Real-time
Flask-SocketIO==5.3.6
python-socketio==5.10.0

# Caching & Background Tasks
redis==5.0.1
celery==5.3.4

# Validation & Serialization
marshmallow==3.21.0
pydantic>=2.0.0

# Utilities
requests==2.31.0
python-dateutil==2.8.2

# Development & Testing
pytest>=8.0.0
pytest-flask==1.3.0
black==23.12.1
flake8==7.0.0 